% Auto-generated: do not edit by hand
\name{''UrlObserver}

\alias{''UrlObserver}

\title{UrlObserver component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''UrlObserver(id=NULL, pathname=NULL)
}

\arguments{
\item{id}{Character. }

\item{pathname}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

